+++ package.json
@@ -6,6 +6,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.10.1",
+    "react-router-dom": "^6.14.1",
     "react-scripts": "5.0.1"
   },
   "scripts": {
